README-tuning: initial treatise
A document on some guiding principles for what the kinematic vs GCode grids are
best used for, and how all the different offsets interact and where they are
stored.

I wrote this up to clarify my thoughts and one insight I had about simplifying
the use of kinematic and XY gcode_offsets vs the use of negative
`positive_min`, as when I went to use one set of macros I got from another
user, which presumably worked for them, it was doing weird things with those
offsets and I ended up with like a +80 Y offset.   This document hopes to
strike a recommendation that trends towards alignment.