Restore pid_k[pid] to be vs. a 255 base
I prefer a 100 base in there, so that the numbers are in units relative to
percentage power output; 255 is sort of weird.  Anyway, the MR to klipper
needs to not change the base without renaming the config field or something.
So, for now, scale the numbers back to be 255ths.